third_party/blink/web_tests/svg/parser/whitespace-angle-invalid-2-expected.txt

CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+ ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ". ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " - ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " + ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " . ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   - ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   + ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   . ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t - ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t + ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t . ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+ ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1grad\f".
This is a testharness.js-based test.
[PASS] Test <angle> invalid value: -
[PASS] Test <angle> invalid value: -deg
[PASS] Test <angle> invalid value: -rad
[PASS] Test <angle> invalid value: -grad
[PASS] Test <angle> invalid value: +
[PASS] Test <angle> invalid value: +deg
[PASS] Test <angle> invalid value: +rad
[PASS] Test <angle> invalid value: +grad
[PASS] Test <angle> invalid value: -.
[PASS] Test <angle> invalid value: -.deg
[PASS] Test <angle> invalid value: -.rad
[PASS] Test <angle> invalid value: -.grad
[PASS] Test <angle> invalid value: .-
[PASS] Test <angle> invalid value: .-deg
[PASS] Test <angle> invalid value: .-rad
[PASS] Test <angle> invalid value: .-grad
[PASS] Test <angle> invalid value: .
[PASS] Test <angle> invalid value: .deg
[PASS] Test <angle> invalid value: .rad
[PASS] Test <angle> invalid value: .grad
[PASS] Test <angle> invalid value: +.
[PASS] Test <angle> invalid value: +.deg
[PASS] Test <angle> invalid value: +.rad
[PASS] Test <angle> invalid value: +.grad
[PASS] Test <angle> invalid value: .E0
[PASS] Test <angle> invalid value: .E0deg
[PASS] Test <angle> invalid value: .E0rad
[PASS] Test <angle> invalid value: .E0grad
[PASS] Test <angle> invalid value: e1
[PASS] Test <angle> invalid value: e1deg
[PASS] Test <angle> invalid value: e1rad
[PASS] Test <angle> invalid value: e1grad
[PASS] Test <angle> invalid value: - 
[PASS] Test <angle> invalid value: -deg 
[PASS] Test <angle> invalid value: -rad 
[PASS] Test <angle> invalid value: -grad 
[PASS] Test <angle> invalid value: + 
[PASS] Test <angle> invalid value: +deg 
[PASS] Test <angle> invalid value: +rad 
[PASS] Test <angle> invalid value: +grad 
[PASS] Test <angle> invalid value: -. 
[PASS] Test <angle> invalid value: -.deg 
[PASS] Test <angle> invalid value: -.rad 
[PASS] Test <angle> invalid value: -.grad 
[PASS] Test <angle> invalid value: .- 
[PASS] Test <angle> invalid value: .-deg 
[PASS] Test <angle> invalid value: .-rad 
[PASS] Test <angle> invalid value: .-grad 
[PASS] Test <angle> invalid value: . 
[PASS] Test <angle> invalid value: .deg 
[PASS] Test <angle> invalid value: .rad 
[PASS] Test <angle> invalid value: .grad 
[PASS] Test <angle> invalid value: +. 
[PASS] Test <angle> invalid value: +.deg 
[PASS] Test <angle> invalid value: +.rad 
[PASS] Test <angle> invalid value: +.grad 
[PASS] Test <angle> invalid value: .E0 
[PASS] Test <angle> invalid value: .E0deg 
[PASS] Test <angle> invalid value: .E0rad 
[PASS] Test <angle> invalid value: .E0grad 
[PASS] Test <angle> invalid value: e1 
[PASS] Test <angle> invalid value: e1deg 
[PASS] Test <angle> invalid value: e1rad 
[PASS] Test <angle> invalid value: e1grad 
[PASS] Test <angle> invalid value: -   
[PASS] Test <angle> invalid value: -deg   
[PASS] Test <angle> invalid value: -rad   
[PASS] Test <angle> invalid value: -grad   
[PASS] Test <angle> invalid value: +   
[PASS] Test <angle> invalid value: +deg   
[PASS] Test <angle> invalid value: +rad   
[PASS] Test <angle> invalid value: +grad   
[PASS] Test <angle> invalid value: -.   
[PASS] Test <angle> invalid value: -.deg   
[PASS] Test <angle> invalid value: -.rad   
[PASS] Test <angle> invalid value: -.grad   
[PASS] Test <angle> invalid value: .-   
[PASS] Test <angle> invalid value: .-deg   
[PASS] Test <angle> invalid value: .-rad   
[PASS] Test <angle> invalid value: .-grad   
[PASS] Test <angle> invalid value: .   
[PASS] Test <angle> invalid value: .deg   
[PASS] Test <angle> invalid value: .rad   
[PASS] Test <angle> invalid value: .grad   
[PASS] Test <angle> invalid value: +.   
[PASS] Test <angle> invalid value: +.deg   
[PASS] Test <angle> invalid value: +.rad   
[PASS] Test <angle> invalid value: +.grad   
[PASS] Test <angle> invalid value: .E0   
[PASS] Test <angle> invalid value: .E0deg   
[PASS] Test <angle> invalid value: .E0rad   
[PASS] Test <angle> invalid value: .E0grad   
[PASS] Test <angle> invalid value: e1   
[PASS] Test <angle> invalid value: e1deg   
[PASS] Test <angle> invalid value: e1rad   
[PASS] Test <angle> invalid value: e1grad   
[PASS] Test <angle> invalid value: -\r\n\t 
[PASS] Test <angle> invalid value: -deg\r\n\t 
[PASS] Test <angle> invalid value: -rad\r\n\t 
[PASS] Test <angle> invalid value: -grad\r\n\t 
[PASS] Test <angle> invalid value: +\r\n\t 
[PASS] Test <angle> invalid value: +deg\r\n\t 
[PASS] Test <angle> invalid value: +rad\r\n\t 
[PASS] Test <angle> invalid value: +grad\r\n\t 
[PASS] Test <angle> invalid value: -.\r\n\t 
[PASS] Test <angle> invalid value: -.deg\r\n\t 
[PASS] Test <angle> invalid value: -.rad\r\n\t 
[PASS] Test <angle> invalid value: -.grad\r\n\t 
[PASS] Test <angle> invalid value: .-\r\n\t 
[PASS] Test <angle> invalid value: .-deg\r\n\t 
[PASS] Test <angle> invalid value: .-rad\r\n\t 
[PASS] Test <angle> invalid value: .-grad\r\n\t 
[PASS] Test <angle> invalid value: .\r\n\t 
[PASS] Test <angle> invalid value: .deg\r\n\t 
[PASS] Test <angle> invalid value: .rad\r\n\t 
[PASS] Test <angle> invalid value: .grad\r\n\t 
[PASS] Test <angle> invalid value: +.\r\n\t 
[PASS] Test <angle> invalid value: +.deg\r\n\t 
[PASS] Test <angle> invalid value: +.rad\r\n\t 
[PASS] Test <angle> invalid value: +.grad\r\n\t 
[PASS] Test <angle> invalid value: .E0\r\n\t 
[PASS] Test <angle> invalid value: .E0deg\r\n\t 
[PASS] Test <angle> invalid value: .E0rad\r\n\t 
[PASS] Test <angle> invalid value: .E0grad\r\n\t 
[PASS] Test <angle> invalid value: e1\r\n\t 
[PASS] Test <angle> invalid value: e1deg\r\n\t 
[PASS] Test <angle> invalid value: e1rad\r\n\t 
[PASS] Test <angle> invalid value: e1grad\r\n\t 
[PASS] Test <angle> invalid value: -\f
[PASS] Test <angle> invalid value: -deg\f
[PASS] Test <angle> invalid value: -rad\f
[PASS] Test <angle> invalid value: -grad\f
[PASS] Test <angle> invalid value: +\f
[PASS] Test <angle> invalid value: +deg\f
[PASS] Test <angle> invalid value: +rad\f
[PASS] Test <angle> invalid value: +grad\f
[PASS] Test <angle> invalid value: -.\f
[PASS] Test <angle> invalid value: -.deg\f
[PASS] Test <angle> invalid value: -.rad\f
[PASS] Test <angle> invalid value: -.grad\f
[PASS] Test <angle> invalid value: .-\f
[PASS] Test <angle> invalid value: .-deg\f
[PASS] Test <angle> invalid value: .-rad\f
[PASS] Test <angle> invalid value: .-grad\f
[PASS] Test <angle> invalid value: .\f
[PASS] Test <angle> invalid value: .deg\f
[PASS] Test <angle> invalid value: .rad\f
[PASS] Test <angle> invalid value: .grad\f
[PASS] Test <angle> invalid value: +.\f
[PASS] Test <angle> invalid value: +.deg\f
[PASS] Test <angle> invalid value: +.rad\f
[PASS] Test <angle> invalid value: +.grad\f
[PASS] Test <angle> invalid value: .E0\f
[PASS] Test <angle> invalid value: .E0deg\f
[PASS] Test <angle> invalid value: .E0rad\f
[PASS] Test <angle> invalid value: .E0grad\f
[PASS] Test <angle> invalid value: e1\f
[PASS] Test <angle> invalid value: e1deg\f
[PASS] Test <angle> invalid value: e1rad\f
[PASS] Test <angle> invalid value: e1grad\f
[PASS] Test <angle> invalid value:  -
[PASS] Test <angle> invalid value:  -deg
[PASS] Test <angle> invalid value:  -rad
[PASS] Test <angle> invalid value:  -grad
[PASS] Test <angle> invalid value:  +
[PASS] Test <angle> invalid value:  +deg
[PASS] Test <angle> invalid value:  +rad
[PASS] Test <angle> invalid value:  +grad
[PASS] Test <angle> invalid value:  -.
[PASS] Test <angle> invalid value:  -.deg
[PASS] Test <angle> invalid value:  -.rad
[PASS] Test <angle> invalid value:  -.grad
[PASS] Test <angle> invalid value:  .-
[PASS] Test <angle> invalid value:  .-deg
[PASS] Test <angle> invalid value:  .-rad
[PASS] Test <angle> invalid value:  .-grad
[PASS] Test <angle> invalid value:  .
[PASS] Test <angle> invalid value:  .deg
[PASS] Test <angle> invalid value:  .rad
[PASS] Test <angle> invalid value:  .grad
[PASS] Test <angle> invalid value:  +.
[PASS] Test <angle> invalid value:  +.deg
[PASS] Test <angle> invalid value:  +.rad
[PASS] Test <angle> invalid value:  +.grad
[PASS] Test <angle> invalid value:  .E0
[PASS] Test <angle> invalid value:  .E0deg
[PASS] Test <angle> invalid value:  .E0rad
[PASS] Test <angle> invalid value:  .E0grad
[PASS] Test <angle> invalid value:  e1
[PASS] Test <angle> invalid value:  e1deg
[PASS] Test <angle> invalid value:  e1rad
[PASS] Test <angle> invalid value:  e1grad
[PASS] Test <angle> invalid value:  - 
[PASS] Test <angle> invalid value:  -deg 
[PASS] Test <angle> invalid value:  -rad 
[PASS] Test <angle> invalid value:  -grad 
[PASS] Test <angle> invalid value:  + 
[PASS] Test <angle> invalid value:  +deg 
[PASS] Test <angle> invalid value:  +rad 
[PASS] Test <angle> invalid value:  +grad 
[PASS] Test <angle> invalid value:  -. 
[PASS] Test <angle> invalid value:  -.deg 
[PASS] Test <angle> invalid value:  -.rad 
[PASS] Test <angle> invalid value:  -.grad 
[PASS] Test <angle> invalid value:  .- 
[PASS] Test <angle> invalid value:  .-deg 
[PASS] Test <angle> invalid value:  .-rad 
[PASS] Test <angle> invalid value:  .-grad 
[PASS] Test <angle> invalid value:  . 
[PASS] Test <angle> invalid value:  .deg 
[PASS] Test <angle> invalid value:  .rad 
[PASS] Test <angle> invalid value:  .grad 
[PASS] Test <angle> invalid value:  +. 
[PASS] Test <angle> invalid value:  +.deg 
[PASS] Test <angle> invalid value:  +.rad 
[PASS] Test <angle> invalid value:  +.grad 
[PASS] Test <angle> invalid value:  .E0 
[PASS] Test <angle> invalid value:  .E0deg 
[PASS] Test <angle> invalid value:  .E0rad 
[PASS] Test <angle> invalid value:  .E0grad 
[PASS] Test <angle> invalid value:  e1 
[PASS] Test <angle> invalid value:  e1deg 
[PASS] Test <angle> invalid value:  e1rad 
[PASS] Test <angle> invalid value:  e1grad 
[PASS] Test <angle> invalid value:  -   
[PASS] Test <angle> invalid value:  -deg   
[PASS] Test <angle> invalid value:  -rad   
[PASS] Test <angle> invalid value:  -grad   
[PASS] Test <angle> invalid value:  +   
[PASS] Test <angle> invalid value:  +deg   
[PASS] Test <angle> invalid value:  +rad   
[PASS] Test <angle> invalid value:  +grad   
[PASS] Test <angle> invalid value:  -.   
[PASS] Test <angle> invalid value:  -.deg   
[PASS] Test <angle> invalid value:  -.rad   
[PASS] Test <angle> invalid value:  -.grad   
[PASS] Test <angle> invalid value:  .-   
[PASS] Test <angle> invalid value:  .-deg   
[PASS] Test <angle> invalid value:  .-rad   
[PASS] Test <angle> invalid value:  .-grad   
[PASS] Test <angle> invalid value:  .   
[PASS] Test <angle> invalid value:  .deg   
[PASS] Test <angle> invalid value:  .rad   
[PASS] Test <angle> invalid value:  .grad   
[PASS] Test <angle> invalid value:  +.   
[PASS] Test <angle> invalid value:  +.deg   
[PASS] Test <angle> invalid value:  +.rad   
[PASS] Test <angle> invalid value:  +.grad   
[PASS] Test <angle> invalid value:  .E0   
[PASS] Test <angle> invalid value:  .E0deg   
[PASS] Test <angle> invalid value:  .E0rad   
[PASS] Test <angle> invalid value:  .E0grad   
[PASS] Test <angle> invalid value:  e1   
[PASS] Test <angle> invalid value:  e1deg   
[PASS] Test <angle> invalid value:  e1rad   
[PASS] Test <angle> invalid value:  e1grad   
[PASS] Test <angle> invalid value:  -\r\n\t 
[PASS] Test <angle> invalid value:  -deg\r\n\t 
[PASS] Test <angle> invalid value:  -rad\r\n\t 
[PASS] Test <angle> invalid value:  -grad\r\n\t 
[PASS] Test <angle> invalid value:  +\r\n\t 
[PASS] Test <angle> invalid value:  +deg\r\n\t 
[PASS] Test <angle> invalid value:  +rad\r\n\t 
[PASS] Test <angle> invalid value:  +grad\r\n\t 
[PASS] Test <angle> invalid value:  -.\r\n\t 
[PASS] Test <angle> invalid value:  -.deg\r\n\t 
[PASS] Test <angle> invalid value:  -.rad\r\n\t 
[PASS] Test <angle> invalid value:  -.grad\r\n\t 
[PASS] Test <angle> invalid value:  .-\r\n\t 
[PASS] Test <angle> invalid value:  .-deg\r\n\t 
[PASS] Test <angle> invalid value:  .-rad\r\n\t 
[PASS] Test <angle> invalid value:  .-grad\r\n\t 
[PASS] Test <angle> invalid value:  .\r\n\t 
[PASS] Test <angle> invalid value:  .deg\r\n\t 
[PASS] Test <angle> invalid value:  .rad\r\n\t 
[PASS] Test <angle> invalid value:  .grad\r\n\t 
[PASS] Test <angle> invalid value:  +.\r\n\t 
[PASS] Test <angle> invalid value:  +.deg\r\n\t 
[PASS] Test <angle> invalid value:  +.rad\r\n\t 
[PASS] Test <angle> invalid value:  +.grad\r\n\t 
[PASS] Test <angle> invalid value:  .E0\r\n\t 
[PASS] Test <angle> invalid value:  .E0deg\r\n\t 
[PASS] Test <angle> invalid value:  .E0rad\r\n\t 
[PASS] Test <angle> invalid value:  .E0grad\r\n\t 
[PASS] Test <angle> invalid value:  e1\r\n\t 
[PASS] Test <angle> invalid value:  e1deg\r\n\t 
[PASS] Test <angle> invalid value:  e1rad\r\n\t 
[PASS] Test <angle> invalid value:  e1grad\r\n\t 
[PASS] Test <angle> invalid value:  -\f
[PASS] Test <angle> invalid value:  -deg\f
[PASS] Test <angle> invalid value:  -rad\f
[PASS] Test <angle> invalid value:  -grad\f
[PASS] Test <angle> invalid value:  +\f
[PASS] Test <angle> invalid value:  +deg\f
[PASS] Test <angle> invalid value:  +rad\f
[PASS] Test <angle> invalid value:  +grad\f
[PASS] Test <angle> invalid value:  -.\f
[PASS] Test <angle> invalid value:  -.deg\f
[PASS] Test <angle> invalid value:  -.rad\f
[PASS] Test <angle> invalid value:  -.grad\f
[PASS] Test <angle> invalid value:  .-\f
[PASS] Test <angle> invalid value:  .-deg\f
[PASS] Test <angle> invalid value:  .-rad\f
[PASS] Test <angle> invalid value:  .-grad\f
[PASS] Test <angle> invalid value:  .\f
[PASS] Test <angle> invalid value:  .deg\f
[PASS] Test <angle> invalid value:  .rad\f
[PASS] Test <angle> invalid value:  .grad\f
[PASS] Test <angle> invalid value:  +.\f
[PASS] Test <angle> invalid value:  +.deg\f
[PASS] Test <angle> invalid value:  +.rad\f
[PASS] Test <angle> invalid value:  +.grad\f
[PASS] Test <angle> invalid value:  .E0\f
[PASS] Test <angle> invalid value:  .E0deg\f
[PASS] Test <angle> invalid value:  .E0rad\f
[PASS] Test <angle> invalid value:  .E0grad\f
[PASS] Test <angle> invalid value:  e1\f
[PASS] Test <angle> invalid value:  e1deg\f
[PASS] Test <angle> invalid value:  e1rad\f
[PASS] Test <angle> invalid value:  e1grad\f
[PASS] Test <angle> invalid value:    -
[PASS] Test <angle> invalid value:    -deg
[PASS] Test <angle> invalid value:    -rad
[PASS] Test <angle> invalid value:    -grad
[PASS] Test <angle> invalid value:    +
[PASS] Test <angle> invalid value:    +deg
[PASS] Test <angle> invalid value:    +rad
[PASS] Test <angle> invalid value:    +grad
[PASS] Test <angle> invalid value:    -.
[PASS] Test <angle> invalid value:    -.deg
[PASS] Test <angle> invalid value:    -.rad
[PASS] Test <angle> invalid value:    -.grad
[PASS] Test <angle> invalid value:    .-
[PASS] Test <angle> invalid value:    .-deg
[PASS] Test <angle> invalid value:    .-rad
[PASS] Test <angle> invalid value:    .-grad
[PASS] Test <angle> invalid value:    .
[PASS] Test <angle> invalid value:    .deg
[PASS] Test <angle> invalid value:    .rad
[PASS] Test <angle> invalid value:    .grad
[PASS] Test <angle> invalid value:    +.
[PASS] Test <angle> invalid value:    +.deg
[PASS] Test <angle> invalid value:    +.rad
[PASS] Test <angle> invalid value:    +.grad
[PASS] Test <angle> invalid value:    .E0
[PASS] Test <angle> invalid value:    .E0deg
[PASS] Test <angle> invalid value:    .E0rad
[PASS] Test <angle> invalid value:    .E0grad
[PASS] Test <angle> invalid value:    e1
[PASS] Test <angle> invalid value:    e1deg
[PASS] Test <angle> invalid value:    e1rad
[PASS] Test <angle> invalid value:    e1grad
[PASS] Test <angle> invalid value:    - 
[PASS] Test <angle> invalid value:    -deg 
[PASS] Test <angle> invalid value:    -rad 
[PASS] Test <angle> invalid value:    -grad 
[PASS] Test <angle> invalid value:    + 
[PASS] Test <angle> invalid value:    +deg 
[PASS] Test <angle> invalid value:    +rad 
[PASS] Test <angle> invalid value:    +grad 
[PASS] Test <angle> invalid value:    -. 
[PASS] Test <angle> invalid value:    -.deg 
[PASS] Test <angle> invalid value:    -.rad 
[PASS] Test <angle> invalid value:    -.grad 
[PASS] Test <angle> invalid value:    .- 
[PASS] Test <angle> invalid value:    .-deg 
[PASS] Test <angle> invalid value:    .-rad 
[PASS] Test <angle> invalid value:    .-grad 
[PASS] Test <angle> invalid value:    . 
[PASS] Test <angle> invalid value:    .deg 
[PASS] Test <angle> invalid value:    .rad 
[PASS] Test <angle> invalid value:    .grad 
[PASS] Test <angle> invalid value:    +. 
[PASS] Test <angle> invalid value:    +.deg 
[PASS] Test <angle> invalid value:    +.rad 
[PASS] Test <angle> invalid value:    +.grad 
[PASS] Test <angle> invalid value:    .E0 
[PASS] Test <angle> invalid value:    .E0deg 
[PASS] Test <angle> invalid value:    .E0rad 
[PASS] Test <angle> invalid value:    .E0grad 
[PASS] Test <angle> invalid value:    e1 
[PASS] Test <angle> invalid value:    e1deg 
[PASS] Test <angle> invalid value:    e1rad 
[PASS] Test <angle> invalid value:    e1grad 
[PASS] Test <angle> invalid value:    -   
[PASS] Test <angle> invalid value:    -deg   
[PASS] Test <angle> invalid value:    -rad   
[PASS] Test <angle> invalid value:    -grad   
[PASS] Test <angle> invalid value:    +   
[PASS] Test <angle> invalid value:    +deg   
[PASS] Test <angle> invalid value:    +rad   
[PASS] Test <angle> invalid value:    +grad   
[PASS] Test <angle> invalid value:    -.   
[PASS] Test <angle> invalid value:    -.deg   
[PASS] Test <angle> invalid value:    -.rad   
[PASS] Test <angle> invalid value:    -.grad   
[PASS] Test <angle> invalid value:    .-   
[PASS] Test <angle> invalid value:    .-deg   
[PASS] Test <angle> invalid value:    .-rad   
[PASS] Test <angle> invalid value:    .-grad   
[PASS] Test <angle> invalid value:    .   
[PASS] Test <angle> invalid value:    .deg   
[PASS] Test <angle> invalid value:    .rad   
[PASS] Test <angle> invalid value:    .grad   
[PASS] Test <angle> invalid value:    +.   
[PASS] Test <angle> invalid value:    +.deg   
[PASS] Test <angle> invalid value:    +.rad   
[PASS] Test <angle> invalid value:    +.grad   
[PASS] Test <angle> invalid value:    .E0   
[PASS] Test <angle> invalid value:    .E0deg   
[PASS] Test <angle> invalid value:    .E0rad   
[PASS] Test <angle> invalid value:    .E0grad   
[PASS] Test <angle> invalid value:    e1   
[PASS] Test <angle> invalid value:    e1deg   
[PASS] Test <angle> invalid value:    e1rad   
[PASS] Test <angle> invalid value:    e1grad   
[PASS] Test <angle> invalid value:    -\r\n\t 
[PASS] Test <angle> invalid value:    -deg\r\n\t 
[PASS] Test <angle> invalid value:    -rad\r\n\t 
[PASS] Test <angle> invalid value:    -grad\r\n\t 
[PASS] Test <angle> invalid value:    +\r\n\t 
[PASS] Test <angle> invalid value:    +deg\r\n\t 
[PASS] Test <angle> invalid value:    +rad\r\n\t 
[PASS] Test <angle> invalid value:    +grad\r\n\t 
[PASS] Test <angle> invalid value:    -.\r\n\t 
[PASS] Test <angle> invalid value:    -.deg\r\n\t 
[PASS] Test <angle> invalid value:    -.rad\r\n\t 
[PASS] Test <angle> invalid value:    -.grad\r\n\t 
[PASS] Test <angle> invalid value:    .-\r\n\t 
[PASS] Test <angle> invalid value:    .-deg\r\n\t 
[PASS] Test <angle> invalid value:    .-rad\r\n\t 
[PASS] Test <angle> invalid value:    .-grad\r\n\t 
[PASS] Test <angle> invalid value:    .\r\n\t 
[PASS] Test <angle> invalid value:    .deg\r\n\t 
[PASS] Test <angle> invalid value:    .rad\r\n\t 
[PASS] Test <angle> invalid value:    .grad\r\n\t 
[PASS] Test <angle> invalid value:    +.\r\n\t 
[PASS] Test <angle> invalid value:    +.deg\r\n\t 
[PASS] Test <angle> invalid value:    +.rad\r\n\t 
[PASS] Test <angle> invalid value:    +.grad\r\n\t 
[PASS] Test <angle> invalid value:    .E0\r\n\t 
[PASS] Test <angle> invalid value:    .E0deg\r\n\t 
[PASS] Test <angle> invalid value:    .E0rad\r\n\t 
[PASS] Test <angle> invalid value:    .E0grad\r\n\t 
[PASS] Test <angle> invalid value:    e1\r\n\t 
[PASS] Test <angle> invalid value:    e1deg\r\n\t 
[PASS] Test <angle> invalid value:    e1rad\r\n\t 
[PASS] Test <angle> invalid value:    e1grad\r\n\t 
[PASS] Test <angle> invalid value:    -\f
[PASS] Test <angle> invalid value:    -deg\f
[PASS] Test <angle> invalid value:    -rad\f
[PASS] Test <angle> invalid value:    -grad\f
[PASS] Test <angle> invalid value:    +\f
[PASS] Test <angle> invalid value:    +deg\f
[PASS] Test <angle> invalid value:    +rad\f
[PASS] Test <angle> invalid value:    +grad\f
[PASS] Test <angle> invalid value:    -.\f
[PASS] Test <angle> invalid value:    -.deg\f
[PASS] Test <angle> invalid value:    -.rad\f
[PASS] Test <angle> invalid value:    -.grad\f
[PASS] Test <angle> invalid value:    .-\f
[PASS] Test <angle> invalid value:    .-deg\f
[PASS] Test <angle> invalid value:    .-rad\f
[PASS] Test <angle> invalid value:    .-grad\f
[PASS] Test <angle> invalid value:    .\f
[PASS] Test <angle> invalid value:    .deg\f
[PASS] Test <angle> invalid value:    .rad\f
[PASS] Test <angle> invalid value:    .grad\f
[PASS] Test <angle> invalid value:    +.\f
[PASS] Test <angle> invalid value:    +.deg\f
[PASS] Test <angle> invalid value:    +.rad\f
[PASS] Test <angle> invalid value:    +.grad\f
[PASS] Test <angle> invalid value:    .E0\f
[PASS] Test <angle> invalid value:    .E0deg\f
[PASS] Test <angle> invalid value:    .E0rad\f
[PASS] Test <angle> invalid value:    .E0grad\f
[PASS] Test <angle> invalid value:    e1\f
[PASS] Test <angle> invalid value:    e1deg\f
[PASS] Test <angle> invalid value:    e1rad\f
[PASS] Test <angle> invalid value:    e1grad\f
[PASS] Test <angle> invalid value: \r\n\t -
[PASS] Test <angle> invalid value: \r\n\t -deg
[PASS] Test <angle> invalid value: \r\n\t -rad
[PASS] Test <angle> invalid value: \r\n\t -grad
[PASS] Test <angle> invalid value: \r\n\t +
[PASS] Test <angle> invalid value: \r\n\t +deg
[PASS] Test <angle> invalid value: \r\n\t +rad
[PASS] Test <angle> invalid value: \r\n\t +grad
[PASS] Test <angle> invalid value: \r\n\t -.
[PASS] Test <angle> invalid value: \r\n\t -.deg
[PASS] Test <angle> invalid value: \r\n\t -.rad
[PASS] Test <angle> invalid value: \r\n\t -.grad
[PASS] Test <angle> invalid value: \r\n\t .-
[PASS] Test <angle> invalid value: \r\n\t .-deg
[PASS] Test <angle> invalid value: \r\n\t .-rad
[PASS] Test <angle> invalid value: \r\n\t .-grad
[PASS] Test <angle> invalid value: \r\n\t .
[PASS] Test <angle> invalid value: \r\n\t .deg
[PASS] Test <angle> invalid value: \r\n\t .rad
[PASS] Test <angle> invalid value: \r\n\t .grad
[PASS] Test <angle> invalid value: \r\n\t +.
[PASS] Test <angle> invalid value: \r\n\t +.deg
[PASS] Test <angle> invalid value: \r\n\t +.rad
[PASS] Test <angle> invalid value: \r\n\t +.grad
[PASS] Test <angle> invalid value: \r\n\t .E0
[PASS] Test <angle> invalid value: \r\n\t .E0deg
[PASS] Test <angle> invalid value: \r\n\t .E0rad
[PASS] Test <angle> invalid value: \r\n\t .E0grad
[PASS] Test <angle> invalid value: \r\n\t e1
[PASS] Test <angle> invalid value: \r\n\t e1deg
[PASS] Test <angle> invalid value: \r\n\t e1rad
[PASS] Test <angle> invalid value: \r\n\t e1grad
[PASS] Test <angle> invalid value: \r\n\t - 
[PASS] Test <angle> invalid value: \r\n\t -deg 
[PASS] Test <angle> invalid value: \r\n\t -rad 
[PASS] Test <angle> invalid value: \r\n\t -grad 
[PASS] Test <angle> invalid value: \r\n\t + 
[PASS] Test <angle> invalid value: \r\n\t +deg 
[PASS] Test <angle> invalid value: \r\n\t +rad 
[PASS] Test <angle> invalid value: \r\n\t +grad 
[PASS] Test <angle> invalid value: \r\n\t -. 
[PASS] Test <angle> invalid value: \r\n\t -.deg 
[PASS] Test <angle> invalid value: \r\n\t -.rad 
[PASS] Test <angle> invalid value: \r\n\t -.grad 
[PASS] Test <angle> invalid value: \r\n\t .- 
[PASS] Test <angle> invalid value: \r\n\t .-deg 
[PASS] Test <angle> invalid value: \r\n\t .-rad 
[PASS] Test <angle> invalid value: \r\n\t .-grad 
[PASS] Test <angle> invalid value: \r\n\t . 
[PASS] Test <angle> invalid value: \r\n\t .deg 
[PASS] Test <angle> invalid value: \r\n\t .rad 
[PASS] Test <angle> invalid value: \r\n\t .grad 
[PASS] Test <angle> invalid value: \r\n\t +. 
[PASS] Test <angle> invalid value: \r\n\t +.deg 
[PASS] Test <angle> invalid value: \r\n\t +.rad 
[PASS] Test <angle> invalid value: \r\n\t +.grad 
[PASS] Test <angle> invalid value: \r\n\t .E0 
[PASS] Test <angle> invalid value: \r\n\t .E0deg 
[PASS] Test <angle> invalid value: \r\n\t .E0rad 
[PASS] Test <angle> invalid value: \r\n\t .E0grad 
[PASS] Test <angle> invalid value: \r\n\t e1 
[PASS] Test <angle> invalid value: \r\n\t e1deg 
[PASS] Test <angle> invalid value: \r\n\t e1rad 
[PASS] Test <angle> invalid value: \r\n\t e1grad 
[PASS] Test <angle> invalid value: \r\n\t -   
[PASS] Test <angle> invalid value: \r\n\t -deg   
[PASS] Test <angle> invalid value: \r\n\t -rad   
[PASS] Test <angle> invalid value: \r\n\t -grad   
[PASS] Test <angle> invalid value: \r\n\t +   
[PASS] Test <angle> invalid value: \r\n\t +deg   
[PASS] Test <angle> invalid value: \r\n\t +rad   
[PASS] Test <angle> invalid value: \r\n\t +grad   
[PASS] Test <angle> invalid value: \r\n\t -.   
[PASS] Test <angle> invalid value: \r\n\t -.deg   
[PASS] Test <angle> invalid value: \r\n\t -.rad   
[PASS] Test <angle> invalid value: \r\n\t -.grad   
[PASS] Test <angle> invalid value: \r\n\t .-   
[PASS] Test <angle> invalid value: \r\n\t .-deg   
[PASS] Test <angle> invalid value: \r\n\t .-rad   
[PASS] Test <angle> invalid value: \r\n\t .-grad   
[PASS] Test <angle> invalid value: \r\n\t .   
[PASS] Test <angle> invalid value: \r\n\t .deg   
[PASS] Test <angle> invalid value: \r\n\t .rad   
[PASS] Test <angle> invalid value: \r\n\t .grad   
[PASS] Test <angle> invalid value: \r\n\t +.   
[PASS] Test <angle> invalid value: \r\n\t +.deg   
[PASS] Test <angle> invalid value: \r\n\t +.rad   
[PASS] Test <angle> invalid value: \r\n\t +.grad   
[PASS] Test <angle> invalid value: \r\n\t .E0   
[PASS] Test <angle> invalid value: \r\n\t .E0deg   
[PASS] Test <angle> invalid value: \r\n\t .E0rad   
[PASS] Test <angle> invalid value: \r\n\t .E0grad   
[PASS] Test <angle> invalid value: \r\n\t e1   
[PASS] Test <angle> invalid value: \r\n\t e1deg   
[PASS] Test <angle> invalid value: \r\n\t e1rad   
[PASS] Test <angle> invalid value: \r\n\t e1grad   
[PASS] Test <angle> invalid value: \r\n\t -\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t -deg\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t -rad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t -grad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t +\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t +deg\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t +rad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t +grad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t -.\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t -.deg\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t -.rad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t -.grad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t .-\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t .-deg\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t .-rad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t .-grad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t .\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t .deg\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t .rad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t .grad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t +.\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t +.deg\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t +.rad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t +.grad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t .E0\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t .E0deg\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t .E0rad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t .E0grad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t e1\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t e1deg\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t e1rad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t e1grad\r\n\t 
[PASS] Test <angle> invalid value: \r\n\t -\f
[PASS] Test <angle> invalid value: \r\n\t -deg\f
[PASS] Test <angle> invalid value: \r\n\t -rad\f
[PASS] Test <angle> invalid value: \r\n\t -grad\f
[PASS] Test <angle> invalid value: \r\n\t +\f
[PASS] Test <angle> invalid value: \r\n\t +deg\f
[PASS] Test <angle> invalid value: \r\n\t +rad\f
[PASS] Test <angle> invalid value: \r\n\t +grad\f
[PASS] Test <angle> invalid value: \r\n\t -.\f
[PASS] Test <angle> invalid value: \r\n\t -.deg\f
[PASS] Test <angle> invalid value: \r\n\t -.rad\f
[PASS] Test <angle> invalid value: \r\n\t -.grad\f
[PASS] Test <angle> invalid value: \r\n\t .-\f
[PASS] Test <angle> invalid value: \r\n\t .-deg\f
[PASS] Test <angle> invalid value: \r\n\t .-rad\f
[PASS] Test <angle> invalid value: \r\n\t .-grad\f
[PASS] Test <angle> invalid value: \r\n\t .\f
[PASS] Test <angle> invalid value: \r\n\t .deg\f
[PASS] Test <angle> invalid value: \r\n\t .rad\f
[PASS] Test <angle> invalid value: \r\n\t .grad\f
[PASS] Test <angle> invalid value: \r\n\t +.\f
[PASS] Test <angle> invalid value: \r\n\t +.deg\f
[PASS] Test <angle> invalid value: \r\n\t +.rad\f
[PASS] Test <angle> invalid value: \r\n\t +.grad\f
[PASS] Test <angle> invalid value: \r\n\t .E0\f
[PASS] Test <angle> invalid value: \r\n\t .E0deg\f
[PASS] Test <angle> invalid value: \r\n\t .E0rad\f
[PASS] Test <angle> invalid value: \r\n\t .E0grad\f
[PASS] Test <angle> invalid value: \r\n\t e1\f
[PASS] Test <angle> invalid value: \r\n\t e1deg\f
[PASS] Test <angle> invalid value: \r\n\t e1rad\f
[PASS] Test <angle> invalid value: \r\n\t e1grad\f
[PASS] Test <angle> invalid value: \f-
[PASS] Test <angle> invalid value: \f-deg
[PASS] Test <angle> invalid value: \f-rad
[PASS] Test <angle> invalid value: \f-grad
[PASS] Test <angle> invalid value: \f+
[PASS] Test <angle> invalid value: \f+deg
[PASS] Test <angle> invalid value: \f+rad
[PASS] Test <angle> invalid value: \f+grad
[PASS] Test <angle> invalid value: \f-.
[PASS] Test <angle> invalid value: \f-.deg
[PASS] Test <angle> invalid value: \f-.rad
[PASS] Test <angle> invalid value: \f-.grad
[PASS] Test <angle> invalid value: \f.-
[PASS] Test <angle> invalid value: \f.-deg
[PASS] Test <angle> invalid value: \f.-rad
[PASS] Test <angle> invalid value: \f.-grad
[PASS] Test <angle> invalid value: \f.
[PASS] Test <angle> invalid value: \f.deg
[PASS] Test <angle> invalid value: \f.rad
[PASS] Test <angle> invalid value: \f.grad
[PASS] Test <angle> invalid value: \f+.
[PASS] Test <angle> invalid value: \f+.deg
[PASS] Test <angle> invalid value: \f+.rad
[PASS] Test <angle> invalid value: \f+.grad
[PASS] Test <angle> invalid value: \f.E0
[PASS] Test <angle> invalid value: \f.E0deg
[PASS] Test <angle> invalid value: \f.E0rad
[PASS] Test <angle> invalid value: \f.E0grad
[PASS] Test <angle> invalid value: \fe1
[PASS] Test <angle> invalid value: \fe1deg
[PASS] Test <angle> invalid value: \fe1rad
[PASS] Test <angle> invalid value: \fe1grad
[PASS] Test <angle> invalid value: \f- 
[PASS] Test <angle> invalid value: \f-deg 
[PASS] Test <angle> invalid value: \f-rad 
[PASS] Test <angle> invalid value: \f-grad 
[PASS] Test <angle> invalid value: \f+ 
[PASS] Test <angle> invalid value: \f+deg 
[PASS] Test <angle> invalid value: \f+rad 
[PASS] Test <angle> invalid value: \f+grad 
[PASS] Test <angle> invalid value: \f-. 
[PASS] Test <angle> invalid value: \f-.deg 
[PASS] Test <angle> invalid value: \f-.rad 
[PASS] Test <angle> invalid value: \f-.grad 
[PASS] Test <angle> invalid value: \f.- 
[PASS] Test <angle> invalid value: \f.-deg 
[PASS] Test <angle> invalid value: \f.-rad 
[PASS] Test <angle> invalid value: \f.-grad 
[PASS] Test <angle> invalid value: \f. 
[PASS] Test <angle> invalid value: \f.deg 
[PASS] Test <angle> invalid value: \f.rad 
[PASS] Test <angle> invalid value: \f.grad 
[PASS] Test <angle> invalid value: \f+. 
[PASS] Test <angle> invalid value: \f+.deg 
[PASS] Test <angle> invalid value: \f+.rad 
[PASS] Test <angle> invalid value: \f+.grad 
[PASS] Test <angle> invalid value: \f.E0 
[PASS] Test <angle> invalid value: \f.E0deg 
[PASS] Test <angle> invalid value: \f.E0rad 
[PASS] Test <angle> invalid value: \f.E0grad 
[PASS] Test <angle> invalid value: \fe1 
[PASS] Test <angle> invalid value: \fe1deg 
[PASS] Test <angle> invalid value: \fe1rad 
[PASS] Test <angle> invalid value: \fe1grad 
[PASS] Test <angle> invalid value: \f-   
[PASS] Test <angle> invalid value: \f-deg   
[PASS] Test <angle> invalid value: \f-rad   
[PASS] Test <angle> invalid value: \f-grad   
[PASS] Test <angle> invalid value: \f+   
[PASS] Test <angle> invalid value: \f+deg   
[PASS] Test <angle> invalid value: \f+rad   
[PASS] Test <angle> invalid value: \f+grad   
[PASS] Test <angle> invalid value: \f-.   
[PASS] Test <angle> invalid value: \f-.deg   
[PASS] Test <angle> invalid value: \f-.rad   
[PASS] Test <angle> invalid value: \f-.grad   
[PASS] Test <angle> invalid value: \f.-   
[PASS] Test <angle> invalid value: \f.-deg   
[PASS] Test <angle> invalid value: \f.-rad   
[PASS] Test <angle> invalid value: \f.-grad   
[PASS] Test <angle> invalid value: \f.   
[PASS] Test <angle> invalid value: \f.deg   
[PASS] Test <angle> invalid value: \f.rad   
[PASS] Test <angle> invalid value: \f.grad   
[PASS] Test <angle> invalid value: \f+.   
[PASS] Test <angle> invalid value: \f+.deg   
[PASS] Test <angle> invalid value: \f+.rad   
[PASS] Test <angle> invalid value: \f+.grad   
[PASS] Test <angle> invalid value: \f.E0   
[PASS] Test <angle> invalid value: \f.E0deg   
[PASS] Test <angle> invalid value: \f.E0rad   
[PASS] Test <angle> invalid value: \f.E0grad   
[PASS] Test <angle> invalid value: \fe1   
[PASS] Test <angle> invalid value: \fe1deg   
[PASS] Test <angle> invalid value: \fe1rad   
[PASS] Test <angle> invalid value: \fe1grad   
[PASS] Test <angle> invalid value: \f-\r\n\t 
[PASS] Test <angle> invalid value: \f-deg\r\n\t 
[PASS] Test <angle> invalid value: \f-rad\r\n\t 
[PASS] Test <angle> invalid value: \f-grad\r\n\t 
[PASS] Test <angle> invalid value: \f+\r\n\t 
[PASS] Test <angle> invalid value: \f+deg\r\n\t 
[PASS] Test <angle> invalid value: \f+rad\r\n\t 
[PASS] Test <angle> invalid value: \f+grad\r\n\t 
[PASS] Test <angle> invalid value: \f-.\r\n\t 
[PASS] Test <angle> invalid value: \f-.deg\r\n\t 
[PASS] Test <angle> invalid value: \f-.rad\r\n\t 
[PASS] Test <angle> invalid value: \f-.grad\r\n\t 
[PASS] Test <angle> invalid value: \f.-\r\n\t 
[PASS] Test <angle> invalid value: \f.-deg\r\n\t 
[PASS] Test <angle> invalid value: \f.-rad\r\n\t 
[PASS] Test <angle> invalid value: \f.-grad\r\n\t 
[PASS] Test <angle> invalid value: \f.\r\n\t 
[PASS] Test <angle> invalid value: \f.deg\r\n\t 
[PASS] Test <angle> invalid value: \f.rad\r\n\t 
[PASS] Test <angle> invalid value: \f.grad\r\n\t 
[PASS] Test <angle> invalid value: \f+.\r\n\t 
[PASS] Test <angle> invalid value: \f+.deg\r\n\t 
[PASS] Test <angle> invalid value: \f+.rad\r\n\t 
[PASS] Test <angle> invalid value: \f+.grad\r\n\t 
[PASS] Test <angle> invalid value: \f.E0\r\n\t 
[PASS] Test <angle> invalid value: \f.E0deg\r\n\t 
[PASS] Test <angle> invalid value: \f.E0rad\r\n\t 
[PASS] Test <angle> invalid value: \f.E0grad\r\n\t 
[PASS] Test <angle> invalid value: \fe1\r\n\t 
[PASS] Test <angle> invalid value: \fe1deg\r\n\t 
[PASS] Test <angle> invalid value: \fe1rad\r\n\t 
[PASS] Test <angle> invalid value: \fe1grad\r\n\t 
[PASS] Test <angle> invalid value: \f-\f
[PASS] Test <angle> invalid value: \f-deg\f
[PASS] Test <angle> invalid value: \f-rad\f
[PASS] Test <angle> invalid value: \f-grad\f
[PASS] Test <angle> invalid value: \f+\f
[PASS] Test <angle> invalid value: \f+deg\f
[PASS] Test <angle> invalid value: \f+rad\f
[PASS] Test <angle> invalid value: \f+grad\f
[PASS] Test <angle> invalid value: \f-.\f
[PASS] Test <angle> invalid value: \f-.deg\f
[PASS] Test <angle> invalid value: \f-.rad\f
[PASS] Test <angle> invalid value: \f-.grad\f
[PASS] Test <angle> invalid value: \f.-\f
[PASS] Test <angle> invalid value: \f.-deg\f
[PASS] Test <angle> invalid value: \f.-rad\f
[PASS] Test <angle> invalid value: \f.-grad\f
[PASS] Test <angle> invalid value: \f.\f
[PASS] Test <angle> invalid value: \f.deg\f
[PASS] Test <angle> invalid value: \f.rad\f
[PASS] Test <angle> invalid value: \f.grad\f
[PASS] Test <angle> invalid value: \f+.\f
[PASS] Test <angle> invalid value: \f+.deg\f
[PASS] Test <angle> invalid value: \f+.rad\f
[PASS] Test <angle> invalid value: \f+.grad\f
[PASS] Test <angle> invalid value: \f.E0\f
[PASS] Test <angle> invalid value: \f.E0deg\f
[PASS] Test <angle> invalid value: \f.E0rad\f
[PASS] Test <angle> invalid value: \f.E0grad\f
[PASS] Test <angle> invalid value: \fe1\f
[PASS] Test <angle> invalid value: \fe1deg\f
[PASS] Test <angle> invalid value: \fe1rad\f
[PASS] Test <angle> invalid value: \fe1grad\f
Harness: the test ran to completion.
